WIP: Improve tile loading for sparse global order reader
This changes the sparse global order reader to fetch tiles in global
order instead of fetching tiles ordered based on fragment id and tile
id.

This net change is an improvement in the efficiency of I/O and  compute
by loading tiles in the order they are operated on.

In the example case we see the number of tiles read drop from 14k to 9k.
This is a result of properly loading tiles when needed them and not
having to purge out tiles due to memory constraints and re-reading them.