SparseGlobalOrderReader: Reuse attribute tiles
This changes the base reader class and the sparse global order reader to
allow for saving and reusing of tiles that have been loaded and
unfilterd but not fully processed. This is most applicable to tiles that
were partially copied to the user buffer but not fully. This change
allows this "last" tile to be reused and the copying to the user buffers
resumed without any repeated I/O or decompression.

The change overall does not have a large effect as this typically
effects only at most one tile per attribute being read.